requirements.txt
flake8
pytest
pytest-mock
pyyaml
requests
